T49: extract TestRunnerUI from ParsekUI
Move test runner window (276 lines, 8 fields, 6 methods) into its own
TestRunnerUI class in UI/. ParsekUI keeps a one-line DrawIfOpen delegate.
Settings window toggle uses testRunnerUI.IsOpen property. Follows the
same pattern as ActionsWindowUI/SpawnControlUI/GroupPickerUI.
